+++ nodejs/test/sha1.js
@@ -10,14 +10,14 @@ function Tests(ASSERT, SHA1, UTIL) {
 
     it('should digest "abc"', function() {
       var md = SHA1.create();
-      md.update('abc');
+      md.update('abc', 'utf8');
       ASSERT.equal(
         md.digest().toHex(), 'a9993e364706816aba3e25717850c26c9cd0d89d');
     });
 
     it('should digest "The quick brown fox jumps over the lazy dog"', function() {
       var md = SHA1.create();
-      md.update('The quick brown fox jumps over the lazy dog');
+      md.update('The quick brown fox jumps over the lazy dog', 'utf8');
       ASSERT.equal(
         md.digest().toHex(), '2fd4e1c67a2d28fced849ee1bb76e7391b93eb12');
     });
@@ -32,8 +32,8 @@ function Tests(ASSERT, SHA1, UTIL) {
     it('should digest "THIS IS A MESSAGE"', function() {
       var md = SHA1.create();
       md.start();
-      md.update('THIS IS ');
-      md.update('A MESSAGE');
+      md.update('THIS IS ', 'utf8');
+      md.update('A MESSAGE', 'utf8');
       // do twice to check continuing digest
       ASSERT.equal(
         md.digest().toHex(), '5f24f4d6499fd2d44df6c6e94be8b14a796c071d');
@@ -44,7 +44,7 @@ function Tests(ASSERT, SHA1, UTIL) {
     it('should digest a long message', function() {
       // Note: might be too slow on old browsers
       var md = SHA1.create();
-      md.update(UTIL.fillString('a', 1000000));
+      md.update(UTIL.fillString('a', 1000000), 'utf8');
       ASSERT.equal(
         md.digest().toHex(), '34aa973cd4c4daa4f61eeb2bdbad27316534016f');
     });